Revised Project
I had to build a messenger application as a project for a design
patterns class. Originally I had planned to reuse and adapt everything in the
original 'Java-IRC' project, but I decided instead to rewrite the entire
project (in a simpler way) because the original code was awful. Like,
really awful.